+++ src/components/modules/Header/Header.tsx
@@ -10,21 +10,16 @@ import HexaLogo from './assets/hexapixel_logo.jpg'
 // CSS Imports
 import StyledHeader from './StyledHeader'
 
-// Types Declarations
-interface HeaderProps {
-  id: string
-}
-
 /**
  * This the Header component.
  */
-const Header = ( { id }: HeaderProps ) => {
+const Header = () => {
   return (
     <StyledHeader>
-      <img src={HexaLogo} alt="Hexapixel Logo" />
+      <img src={HexaLogo} alt='Hexapixel Logo' />
       <nav>
-        <Link to="/">Home</Link>
-        <Link to="resume">Resume</Link>
+        <Link to='/'>Home</Link>
+        <Link to='resume'>Resume</Link>
       </nav>
     </StyledHeader>
   )