+++ src/components/pages/Home/StyledHome.ts
@@ -10,9 +10,9 @@ import hexarepeat from 'globalAssets/images/bgrep.png'
 
 const StyledHome = styled.section`
 /* MOBILE FIRST DEFAULT STYLES */
+  border-bottom: 1px solid ${defaultColors.cyberBlue};
   background-image: url(${hexarepeat});
   background-repeat: repeat;
-  border-bottom: 1px solid ${defaultColors.cyberBlue};
   max-width: 100%;
 
 /* Min width of 375 */
@@ -27,7 +27,6 @@ const StyledHome = styled.section`
 
 /* Min width of 576 */
   ${media.landscapePhones `
-    
   `}
 
 /* Min width of 768 */
@@ -37,7 +36,6 @@ const StyledHome = styled.section`
 
 /* Min width of 1024 */
   ${media.desktop `
-    
   `}
 
 /* Min width of 1200 */