+++ src/globalAssets/data/education.ts
@@ -1,0 +1,21 @@
+/**
+ * Define data for the education section.
+ */
+const education = [
+  {
+    dates: 'APR 2019 - JUNE 2019',
+    desc: 'The Web Development Bootcamp is a 12 week, 40+ hours/week bootcamp. It covers the most popular and marketable languages and frameworks used in Web Development today.',
+    location: 'Lighthouse Labs, Toronto',
+    program: 'Web Development Bootcamp',
+    topics: 'The topics covered are: JavaScript, Node.js, Express, PostgreSQL, MongoDB, React, Ruby on Rails, Automated Testing, and Computer Science Fundamentals.'
+  },
+  {
+    dates: 'June 2010',
+    desc: 'Digital Media production based in Graphic Design.',
+    location: 'EDMONTON ARTS COLLEGE',
+    program: 'DIGITAL MEDIA PRODUCTION PROGRAM',
+    topics: 'Topics Covered: Education, Student Instruction, Front End Development (HTML5, CSS3), JavaScript, jQuery, Wordpress, Drupal, Joomla!, Design Interpretation, Career Guidence'
+  }
+]
+
+export default education