+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2023-03-31
+
+### Added
+  - Proper testing to AccordionItem component
+  - Proper testing to Project component
+  - Proper testing of AccordionWrap module
+
+### Changed
+  - Edited Eslint, Jest, and TS configuration to work properly.
+  - Created columns for Projects and Skills section on bigger sized screens.
+  - Updated Portfolio Data
+  - Cleaned up Resume Styles
+  - Exported out Education data to a file
+
 ## [0.1.1] - 2023-03-28
 
 ### Added